Converting Movement variables to double

+++ Rain/src/com/aadi/rain/Game.java
@@ -135,10 +135,10 @@ public class Game extends Canvas implements Runnable {
 
 		screen.clear();
 
-		int xScroll = player.getX() - screen.width / 2;
-		int yScroll = player.getY() - screen.height / 2;
+		double xScroll = player.getX() - screen.width / 2;
+		double yScroll = player.getY() - screen.height / 2;
 
-		level.render(xScroll, yScroll, screen);
+		level.render((int) xScroll, (int) yScroll, screen);
 
 		//screen.renderSheet(40, 40, SpriteSheet.player_down, false);
